### 2025-06-12 #8395 @editorialbot
[PRE REVIEW]: SpecialFunctions: A C# package of special functions for scientific computing with MATLAB-compatible API
**Submitting author:** @jake-w-liu (<a href="http://orcid.org/0000-0001-5458-7917">Jake W. Liu</a>)
**Repository:** https://github.com/jake-w-liu/SpecialFunctions
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.1
