### 2025-06-14 #8395 @jake-w-liu
Hi @danielskatz , thank you for your feedback!

- I’ve added the missing MIT License file to the repository and removed the ambiguous phrase "or as applicable" from the README for clarity.

- The phrase "built upon Math.NET Numerics and AminKH's Legendre Polynomials repository" has been revised in the paper.md for specificity. The package reuses core numerical routines from [Math.NET Numerics](https://github.com/mathnet/mathnet-numerics) for functions like Bessel and error functions, and adapts Legendre polynomial implementations from [AminKH’s repository](https://github.com/AminKH/Legendre-Polynomials), modifying the API to match MATLAB conventions.

- I’ve corrected the date field format in paper.md as per JOSS requirements.

Please let me know if any further changes are needed. Thanks again!
