### 2025-06-15 #8395 @danielskatz
@jake-w-liu - Is any of the code in your repository from these sources?

Again, I'm trying to understand how to identify (in the repo) code that you have created vs code that you have used.
