### 2025-06-16 #8395 @jake-w-liu
@danielskatz  Apologies for the confusion, and thank you for your patience. Let me clarify, in my repo:

- The file [Amos.cs](https://github.com/jake-w-liu/SpecialFunctions/blob/main/src/Amos.cs) contains code derived from [Math.NET Numerics’ adaption of the Amos module](https://github.com/mathnet/mathnet-numerics/tree/master/src/Numerics/SpecialFunctions/Amos), particularly for the implementation of Bessel-related functions.

- The same file also includes additional functions such as gamma, error functions, and related routines, which are my own contributions. I’ve grouped them under the Amos class for organizational consistency, but they are not present in the original sources.

- [Polynomial.cs](https://github.com/jake-w-liu/SpecialFunctions/blob/main/src/Polynomials.cs) is a subset of [Math.NET Numerics' polynomial implementation](https://github.com/mathnet/mathnet-numerics/blob/master/src/Numerics/Polynomial.cs), which is used for the evaluation of Polynomials for special function computation.

- The file [SpecialFunctions.cs](https://github.com/jake-w-liu/SpecialFunctions/blob/main/src/SpecialFunctions.cs) consists primarily of my own code. This includes the design of a unified API layer that wraps lower-level functions, exposing them in a MATLAB-compatible form.

- Within that file, the Legendre polynomial routines are adapted from [AminKH’s repository](https://github.com/AminKH/Legendre-Polynomials), with modifications to MATLAB conventions and to support normalized modes and to align with the overall API style.

I hope this helps clarify the origin of the components and my contributions. Please let me know if further detail is needed.
