### 2025-06-20 #8395 @lockwo
Hi 👋 @Aljanabi1989, @RajashreeMR, @JeWaVe, @jasonyaaa,  would two (or more) of you be willing to review this submission for JOSS?
