### 2025-06-26 #8395 @lockwo
Hi 👋 @ruchithams, @Codewiper, @gokhansatilmis, @LyceanEM, would two of you be willing to review this submission for JOSS?
