### 2025-06-26 #8395 @JeWaVe
Hi. Yes I can. I'll have a look these days.

Envoyé depuis Proton Mail pour Android

-------- Message d'origine --------
Le 20/06/2025 23:18, Owen L  a écrit :

> lockwo left a comment [(openjournals/joss-reviews#8395)](https://github.com/openjournals/joss-reviews/issues/8395#issuecomment-2992965725)
>
> Hi 👋 ***@***.***(https://github.com/Aljanabi1989), ***@***.***(https://github.com/RajashreeMR), ***@***.***(https://github.com/JeWaVe), ***@***.***(https://github.com/jasonyaaa), would two (or more) of you be willing to review this submission for JOSS?
>
