### 2025-07-03 86025a6
remove obj and bin + add dotnet test to CI
