+++ obj/Debug/net9.0/SpecialFunctions.AssemblyInfo.cs
@@ -1,22 +1,0 @@
-//------------------------------------------------------------------------------
-// <auto-generated>
-//     This code was generated by a tool.
-//
-//     Changes to this file may cause incorrect behavior and will be lost if
-//     the code is regenerated.
-// </auto-generated>
-//------------------------------------------------------------------------------
-
-using System;
-using System.Reflection;
-
-[assembly: System.Reflection.AssemblyCompanyAttribute("SpecialFunctions")]
-[assembly: System.Reflection.AssemblyConfigurationAttribute("Debug")]
-[assembly: System.Reflection.AssemblyFileVersionAttribute("1.0.1.0")]
-[assembly: System.Reflection.AssemblyInformationalVersionAttribute("1.0.1+280fa747eb82ee2ed03b7cdc4efd3338ea809a68")]
-[assembly: System.Reflection.AssemblyProductAttribute("SpecialFunctions")]
-[assembly: System.Reflection.AssemblyTitleAttribute("SpecialFunctions")]
-[assembly: System.Reflection.AssemblyVersionAttribute("1.0.1.0")]
-
-// Generated by the MSBuild WriteCodeFragment class.
-
