+++ obj/Release/net481/SpecialFunctions.csproj.FileListAbsolute.txt
@@ -1,143 +1,0 @@
-/Users/jake/SpecialFunctions/bin/Release/net481/Microsoft.TestPlatform.AdapterUtilities.dll
-/Users/jake/SpecialFunctions/bin/Release/net481/Microsoft.VisualStudio.TestPlatform.MSTest.TestAdapter.dll
-/Users/jake/SpecialFunctions/bin/Release/net481/Microsoft.VisualStudio.TestPlatform.MSTestAdapter.PlatformServices.dll
-/Users/jake/SpecialFunctions/bin/Release/net481/SpecialFunctions.dll.config
-/Users/jake/SpecialFunctions/bin/Release/net481/SpecialFunctions.dll
-/Users/jake/SpecialFunctions/bin/Release/net481/SpecialFunctions.pdb
-/Users/jake/SpecialFunctions/bin/Release/net481/Microsoft.ApplicationInsights.dll
-/Users/jake/SpecialFunctions/bin/Release/net481/Microsoft.VisualStudio.CodeCoverage.Shim.dll
-/Users/jake/SpecialFunctions/bin/Release/net481/Microsoft.Testing.Extensions.Telemetry.dll
-/Users/jake/SpecialFunctions/bin/Release/net481/Microsoft.Testing.Extensions.TrxReport.Abstractions.dll
-/Users/jake/SpecialFunctions/bin/Release/net481/Microsoft.Testing.Extensions.VSTestBridge.dll
-/Users/jake/SpecialFunctions/bin/Release/net481/Microsoft.Testing.Platform.dll
-/Users/jake/SpecialFunctions/bin/Release/net481/Microsoft.Testing.Extensions.MSBuild.dll
-/Users/jake/SpecialFunctions/bin/Release/net481/Microsoft.TestPlatform.CoreUtilities.dll
-/Users/jake/SpecialFunctions/bin/Release/net481/Microsoft.TestPlatform.PlatformAbstractions.dll
-/Users/jake/SpecialFunctions/bin/Release/net481/Microsoft.VisualStudio.TestPlatform.ObjectModel.dll
-/Users/jake/SpecialFunctions/bin/Release/net481/Microsoft.VisualStudio.TestPlatform.TestFramework.Extensions.dll
-/Users/jake/SpecialFunctions/bin/Release/net481/Microsoft.VisualStudio.TestPlatform.TestFramework.dll
-/Users/jake/SpecialFunctions/bin/Release/net481/System.Buffers.dll
-/Users/jake/SpecialFunctions/bin/Release/net481/System.Collections.Immutable.dll
-/Users/jake/SpecialFunctions/bin/Release/net481/System.Diagnostics.DiagnosticSource.dll
-/Users/jake/SpecialFunctions/bin/Release/net481/System.Memory.dll
-/Users/jake/SpecialFunctions/bin/Release/net481/System.Numerics.Vectors.dll
-/Users/jake/SpecialFunctions/bin/Release/net481/System.Reflection.Metadata.dll
-/Users/jake/SpecialFunctions/bin/Release/net481/System.Runtime.CompilerServices.Unsafe.dll
-/Users/jake/SpecialFunctions/bin/Release/net481/System.Threading.Tasks.Extensions.dll
-/Users/jake/SpecialFunctions/bin/Release/net481/cs/Microsoft.Testing.Extensions.Telemetry.resources.dll
-/Users/jake/SpecialFunctions/bin/Release/net481/de/Microsoft.Testing.Extensions.Telemetry.resources.dll
-/Users/jake/SpecialFunctions/bin/Release/net481/es/Microsoft.Testing.Extensions.Telemetry.resources.dll
[…]
