### 2025-07-03 eba3885
run tests only on LTS dotnet
