### 2025-07-03 f7ced02
fix ci for tests
