+++ README.md
@@ -1,4 +1,3 @@
-
 # SpecialFunctions
 
 A minimal C# library providing numerical implementations of common special functions, such as:
@@ -73,6 +72,42 @@ Complex bessel = besselj(1.5, new Complex(1.0, 0.5));
 ```
 
 For detailed functionalities, please visit the [manual page](./docs/MANUAL.md).
+
+## Contributing
+
+We welcome contributions from the community. Here are some ways you can contribute to the project.
+
+### Reporting Issues
+
+If you find a bug or have a problem with the software, please open an issue on our GitHub repository.
+
+When reporting an issue, please include the following information:
+
+*   A clear and descriptive title.
+*   A detailed description of the issue, including steps to reproduce it.
+*   The version of the software you are using.
+*   Any relevant error messages or logs.
+
+### Contributing Code
+
+If you would like to contribute code to the project, please follow these steps:
+
+1.  Fork the repository.
+2.  Create a new branch for your feature or bug fix.
+3.  Make your changes and commit them with a clear and descriptive commit message.
+4.  Push your changes to your fork.
+5.  Open a pull request to the main repository.
+
+Please ensure that your code adheres to the existing coding style and that all tests pass.
+
+### Seeking Support
+
+If you have a question or need help with the software, you can:
+
+*   Open an issue on our GitHub repository.
+*   Contact the maintainers directly.
+
+We will do our best to respond to your request in a timely manner.
 
 ## Notes
 
