### 2025-07-05 ae25059
remove bin/obj
