+++ obj/Debug/net481/SpecialFunctions.AssemblyInfo.cs
@@ -1,22 +1,0 @@
-//------------------------------------------------------------------------------
-// <auto-generated>
-//     This code was generated by a tool.
-//
-//     Changes to this file may cause incorrect behavior and will be lost if
-//     the code is regenerated.
-// </auto-generated>
-//------------------------------------------------------------------------------
-
-using System;
-using System.Reflection;
-
-[assembly: System.Reflection.AssemblyCompanyAttribute("SpecialFunctions")]
-[assembly: System.Reflection.AssemblyConfigurationAttribute("Debug")]
-[assembly: System.Reflection.AssemblyFileVersionAttribute("1.0.1.0")]
-[assembly: System.Reflection.AssemblyInformationalVersionAttribute("1.0.1+a58ee95d6ca9c009852773135f149d611d32b31a")]
-[assembly: System.Reflection.AssemblyProductAttribute("SpecialFunctions")]
-[assembly: System.Reflection.AssemblyTitleAttribute("SpecialFunctions")]
-[assembly: System.Reflection.AssemblyVersionAttribute("1.0.1.0")]
-
-// Generated by the MSBuild WriteCodeFragment class.
-
