+++ obj/Debug/net5.0/SpecialFunctions.AssemblyInfo.cs
@@ -1,22 +1,0 @@
-//------------------------------------------------------------------------------
-// <auto-generated>
-//     This code was generated by a tool.
-//
-//     Changes to this file may cause incorrect behavior and will be lost if
-//     the code is regenerated.
-// </auto-generated>
-//------------------------------------------------------------------------------
-
-using System;
-using System.Reflection;
-
-[assembly: System.Reflection.AssemblyCompanyAttribute("SpecialFunctions")]
-[assembly: System.Reflection.AssemblyConfigurationAttribute("Debug")]
-[assembly: System.Reflection.AssemblyFileVersionAttribute("1.0.1.0")]
-[assembly: System.Reflection.AssemblyInformationalVersionAttribute("1.0.1+0105273ab6274b4af9a4717d831928a6d9abc549")]
-[assembly: System.Reflection.AssemblyProductAttribute("SpecialFunctions")]
-[assembly: System.Reflection.AssemblyTitleAttribute("SpecialFunctions")]
-[assembly: System.Reflection.AssemblyVersionAttribute("1.0.1.0")]
-
-// Generated by the MSBuild WriteCodeFragment class.
-
