### 2025-07-05 07a8ca8
docs: add installation instructions and supported frameworks
