### 2025-07-06 4a0db77
docs: update build and test section
