### 2025-07-06 314f40a
ci: add workflow to run tests
