### 2025-07-06 #8516 @jake-w-liu
Hi @JeWaVe ,

Thank you so much for your thorough review and your very kind and encouraging words. I'm delighted that you see the value in this work and found the submission to be of high quality.
I have addressed the five issues you opened in the repository—thank you for your keen eye in spotting those areas for improvement!

Regarding your other comments:
  > [EDIT] code looks like reusing code from other repositories. I'm not versed into licensing subtleties but it should be checked by someone who is.

This is an excellent and important point. You are correct; the library intentionally builds upon established, high-quality open-source code. I've been careful to ensure all licensing is compatible. Specifically:
 * The core numerical routines are adapted from Math.NET Numerics, which is distributed under the MIT License.
* The Legendre polynomial implementations are adapted from AminKH's Legendre-Polynomials repository, with notable modifications.

Since SpecialFunctions is also licensed under the MIT License, it should be fully compatible with these dependencies. I have included acknowledgements in both the README.md and the paper to give credit to these foundational projects. Thank you for ensuring this was properly checked.

  > [EDIT 2] It might be worth to mention that it's possible to p/invoke c++ stdlib (since c++17) or boost but that is a pain compared to a plain c# implementation

 This is an excellent point, and thank you for raising it. You're right that mentioning P/Invoke as an alternative helps to better frame the"Statement of Need." While calling native C++ libraries is certainly possible, it introduces significant friction regarding build complexity and cross-platform portability. SpecialFunctions is designed to avoid precisely this "pain" by providing a pure, idiomatic C# solution. A key advantage of this approach, combined with the permissive MIT license, is that it gives developers maximum flexibility: they can either consume the library as a standard NuGet package or, for greater control, include the source files directly into their own projects—an option not feasible with libraries under more restrictive licenses.

Thank you again for your time and your constructive feedback, which has definitely improved the project. Please let me know if there is anything else I can address!
