### 2025-07-27 #8516 @jasonyaaa
Hi @jake-w-liu

I have reviewed the updated SpecialFunctions package and its accompanying manuscript in detail and am pleased to confirm that the submission fully satisfies JOSS’s requirements with no outstanding concerns. The software installs seamlessly via standard .NET tooling, is supported by MATLAB-style API documentation and practical usage examples, and is fortified by a comprehensive automated test suite integrated into continuous integration workflows. The paper succinctly motivates the need for a permissively licensed C# special-functions library, situates it within the broader numerical-computing ecosystem, and communicates its technical content with clarity and precision for both specialist and interdisciplinary audiences. All citations are accurate and correctly formatted, and the writing flows smoothly throughout. In short, the authors have delivered a polished, high-quality submission that is unequivocally ready for publication.
