### 2025-07-27 #8516 @jasonyaaa
Dear @lockwo,
Thank you for checking in. I have no further questions or concerns—the package and paper fully meet JOSS requirements and are ready for publication.
