### 2025-09-04 #8516 @jake-w-liu
Hello @lockwo,

I’ve updated the paper in the repository to include my ORCID number.

Regarding the DOI for Handbook of Mathematical Functions, the currently listed DOI 10.2307/1266136 appears to be incorrect, as it refers to a review of the book rather than the book itself.

Thank you for your help!
