### 2025-09-05 c8bc19d
fix typo nrom -> norm
