### 2025-09-05 #8516 @jake-w-liu
Hi @lockwo,

thank you for the detailed feedback!

- The "nrom" was indeed a typo — it should have been "norm". A stupid typo 🫣 I’ve corrected this throughout the code.

- I’ve made a new tagged release at v1.0.2, and updated the paper accordingly.

- The over-complete API listing has been consolidated into function families in the paper.

- The .DS_Store has been removed from the repository and added to .gitignore.

The archive DOI for the v1.0.2 release is: [10.5281/zenodo.17060640](https://doi.org/10.5281/zenodo.17060640).

Thanks again for your help in the review process!
