### 2025-09-09 #8516 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
